Testin pull request. Want to make sure changes are squashed (#1)
* Refactoring app structure

Moved files arround. Everything compiles and run. However styles are not applied yet

* Refactgoring app structure - done

Css was not working because I forgot to remove some module.ids and absolute paths

Modification during PullRequest that squashes everything together. Hopefully this will make git think that the additions in one commit and the deletions in the other are renames (they were, but comited separately)